Modularize CloudSync credential routes
Reduce the route façade to composition, separate identity, projection, and token responsibilities, and group integration scenarios by behavior.
